Remove dialog component and move other existing components into the components folder / module. The components repository will also have some UI related behaviors and services, so we want to have `components`, `behaviors` and `services` as top level folders.
