+++ source/components/button/button.ts
@@ -1,11 +1,17 @@
-﻿// uses utilities.d.ts
-// uses typings/angularjs
+﻿// uses typings/angularj
 // uses typings/lodash
+// uses typescript-angular-utilities
 
-module rl.components.button {
+// /// <reference path='../../../typings/angularjs/angular.d.ts' />
+// /// <reference path="../../../typings/lodash/lodash.d.ts" />
+// /// <reference path='../../../libraries/typescript-angular-utilities/typings/utility.d.ts' />
+
+module rl.ui.components.button {
 	'use strict';
+
+	import __promiseUtility = rl.utilities.services.promise;
 	
-	export var moduleName: string = 'rl.components.button';
+	export var moduleName: string = 'rl.ui.components.button';
 	
 	export var directiveName: string = 'rlButton';
 	export var controllerName: string = 'ButtonController';
@@ -14,17 +20,19 @@ module rl.components.button {
 		busy: boolean;
 		action(...params: any[]): ng.IPromise<any>;
 		action(...params: any[]): void;
+		size: string;
 	}
 	
 	export interface IButtonController {
 		busy: boolean;
 		trigger(): void;
 	}
-
-	export class ButtonController {
-		static $inject: string[] = ['$scope',  utilities.promise.serviceName];
-		constructor(private $scope: IButtonScope, private promiseUtility: utilities.promise.IPromiseUtility) {
+	
+	class ButtonController {
+		static $inject: string[] = ['$scope', __promiseUtility.serviceName];
+		constructor(private $scope: IButtonScope, private promiseUtility: __promiseUtility.IPromiseUtility) {
 			this.busy = $scope.busy;
+			this.sizeClass = $scope.size != null ? 'btn-' + $scope.size : null;
 	
 			if (!_.isUndefined($scope.busy)) {
 				$scope.$watch('busy', (value: boolean): void => {
@@ -55,9 +63,10 @@ module rl.components.button {
 		}
 	
 		busy: boolean;
+		sizeClass: string;
 	}
 	
-	export function button(): ng.IDirective {
+	function button(): angular.IDirective {
 		'use strict';
 		return {
 			restrict: 'E',
@@ -69,13 +78,14 @@ module rl.components.button {
 				type: '@',
 				ngDisabled: '=',
 				buttonRightAligned: '=',
+				size: '@',
 			},
 			controller: controllerName,
 			controllerAs: 'button',
 		};
 	}
 	
-	angular.module(moduleName, [utilities.promise.moduleName])
+	angular.module(moduleName, [__promiseUtility.moduleName])
 		.directive(directiveName, button)
 		.controller(controllerName, ButtonController);
 }
