+++ source/components/userRating/userRating.html
