--- source/dialog/dialog.html
@@ -1,5 +1,0 @@
-﻿<div class="modal fade" role="dialog">
-	<div class="modal-dialog {{dialogClass}}">
-		<div ng-transclude></div>
-	</div>
-</div>
