+++ source/services/services.module.ts
@@ -1,0 +1,9 @@
+// /// <reference path='../../typings/angularjs/angular.d.ts' />
+
+module rl.ui.services {
+	export var moduleName: string = 'rl.ui.services';
+	
+	angular.module(moduleName, [
+		
+	]);
+}
