+++ source/ui.module.ts
@@ -1,0 +1,13 @@
+// uses angularjs
+
+/// <reference path='components/components.module.ts' />
+/// <reference path='services/services.module.ts' />
+
+module rl.ui {
+	export var moduleName: string = 'rl.ui';
+
+	angular.module(name, [
+		components.moduleName,
+		services.moduleName,
+	]);
+}
